+++ examples/billboard.rs
@@ -75,7 +75,7 @@ fn setup(
             dimension: ShapeDimension::Volume,
         })
         .render(ParticleTextureModifier {
-            texture: texture_handle.clone(),
+            texture: texture_handle,
         })
         .render(BillboardModifier {})
         .render(ColorOverLifetimeModifier { gradient })